Moved Registerable into an Interfaces folder.
